NotificationChannel on server side

+++ GemsJaxServer/src/org/gemsjax/server/communication/channel/handler/RequestChannelHandler.java
@@ -4,8 +4,8 @@ import org.gemsjax.server.module.OnlineUser;
 
 public interface RequestChannelHandler {
 
-	public void onAcceptRequest(OnlineUser user, int requestId, String referenceId);
-	public void onRejectRequest(OnlineUser user, int requestId, String referenceId);
+	public void onAcceptRequest(OnlineUser user, long requestId, String referenceId);
+	public void onRejectRequest(OnlineUser user, long requestId, String referenceId);
 	
 	public void onGetAllRequests(OnlineUser user, String referenceId);
 }
